Added highlight element to markup
