+++ MacroPepelelipa/MacroPepelelipa/Helpers/MarkupKit/Common/Elements/Escaping/MarkdownCodeEscaping.swift
@@ -27,7 +27,7 @@ open class MarkdownCodeEscaping: MarkdownElement {
       .map { (value: UInt16) -> String in String(format: "%04x", value) }
       .reduce("") { (string: String, character: String) -> String in
         return "\(string)\(character)"
-    }
+      }
     attributedString.replaceCharacters(in: range, with: escapedString)
   }
