+++ MacroPepelelipa/MacroPepelelipa/Helpers/MarkupKit/Common/Elements/Escaping/MarkdownUnescaping.swift
@@ -22,7 +22,9 @@ open class MarkdownUnescaping: MarkdownElement {
   open func match(_ match: NSTextCheckingResult, attributedString: NSMutableAttributedString) {
     let range = NSRange(location: match.range.location + 1, length: 4)
     let matchString = attributedString.attributedSubstring(from: range).string
-    guard let unescapedString = matchString.unescapeUTF16() else { return }
+    guard let unescapedString = matchString.unescapeUTF16() else {
+        return
+    }
     attributedString.replaceCharacters(in: match.range, with: unescapedString)
   }
 }
