+++ MacroPepelelipa/MacroPepelelipa/Helpers/MarkupKit/Common/Protocols/MarkdownCommonElement.swift
