+++ MacroPepelelipa/MacroPepelelipa/Helpers/MarkupKit/Common/Protocols/MarkdownElement.swift
