+++ MacroPepelelipa/MacroPepelelipa/Helpers/MarkupKit/Common/Protocols/MarkdownStyle.swift
